Refactor: 복수명의 파일들 단수명 처리
