+++ controllers/book-detail.controller.js
@@ -1,0 +1,1 @@
+const bookDetailService = require('../services/book-detail.services');
