+++ routes/review.router.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 
-const reviewController = require('../controllers/review.controllers');
+const reviewController = require('../controllers/review.controller');
 
 router.post('', reviewController.createReview);
 router.patch('', reviewController.updateReview);
